Safari-609.1.11.1/LayoutTests/media/modern-media-controls/scrubber-support/scrubber-support-click-expected.txt

Testing the ScrubberSupport behavior when clicking the scrubber.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scrubber.getBoundingClientRect().width became different from 0
PASS media.currentTime is within 0.2 of 3.1872
PASS successfullyParsed is true

TEST COMPLETE
